requirements.txt
torch
torchvision
transformers
pillow
numpy
novelAI
novelai-sdk
